Creating separate component in call booking and moving label to a hook
