+++ src/components/CallBooking/CallBooking.tsx
@@ -3,7 +3,6 @@ import { useTranslation } from 'react-i18next'
 
 import { type CallBooking as CallBookingData, CallBookingPurpose, CallBookingType } from '@schemas/callBooking'
 import { DateFormat } from '@utils/i18n/date/patterns'
-import { tPlural } from '@utils/i18n/plural'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { Button } from '@ui/Button/Button'
 import { LinkButton } from '@ui/Button/LinkButton'
@@ -15,14 +14,12 @@ import { Container } from '@components/Container/Container'
 
 import './callBooking.scss'
 
+import { useCallBookingCountdownLabel } from './useCallBookingCountdownLabel'
+
 export interface CallBookingProps {
   callBooking?: CallBookingData
   onBookCall?: () => void
 }
-
-type CountdownDescriptor =
-  | { kind: 'days' | 'hours', value: number }
-  | { kind: 'startingSoon' | 'none' }
 
 const EmptyState = ({ onBookCall }: { onBookCall?: () => void }) => {
   const { t } = useTranslation()
@@ -42,34 +39,55 @@ const EmptyState = ({ onBookCall }: { onBookCall?: () => void }) => {
   )
 }
 
-const getCountdownDescriptor = (now: number, startAt: Date): CountdownDescriptor => {
-  const startTime = startAt.getTime()
+const OnboardingCallCoverage = () => {
+  const { t } = useTranslation()
+  const whatWeWillCoverItems: Array<[string, string]> = [
+    ['business_and_books', t('callBookings:label.cover_business_and_books', 'Walk through your business and books')],
+    ['accounts_and_documents', t('callBookings:label.cover_accounts_and_documents', 'Connect your accounts and documents')],
+    ['first_month_expectations', t('callBookings:label.cover_first_month_expectations', 'Set expectations for our first month')],
+  ]
 
-  if (now >= startTime) {
-    return { kind: 'none' }
-  }
+  return (
+    <>
+      <Span nonAria className='Layer__CallBooking__Divider' />
 
-  const millisecondsUntilStart = startTime - now
-
-  if (millisecondsUntilStart < 60 * 60 * 1000) {
-    return { kind: 'startingSoon' }
-  }
-
-  const hoursUntilStart = millisecondsUntilStart / (60 * 60 * 1000)
-
-  if (hoursUntilStart < 24) {
-    return { kind: 'hours', value: Math.floor(hoursUntilStart) }
-  }
-
-  return {
-    kind: 'days',
-    value: Math.floor(millisecondsUntilStart / (24 * 60 * 60 * 1000)),
-  }
+      <VStack pbe='md'>
+        <Span
+          nonAria
+          size='2xs'
+          variant='subtle'
+          pbe='sm'
+          className='Layer__CallBooking__CoverageEyebrow'
+        >
+          {t('callBookings:label.what_well_cover', 'What we\'ll cover')}
+        </Span>
+        <VStack role='list' gap='xs'>
+          {whatWeWillCoverItems.map(([key, label]) => (
+            <HStack
+              key={key}
+              className='Layer__CallBooking__CoverageItem'
+              align='start'
+              gap='sm'
+              role='listitem'
+            >
+              <Span nonAria className='Layer__CallBooking__CoverageBadge'>
+                <Check size={12} strokeWidth={2.5} />
+              </Span>
+              <Span size='sm' className='Layer__CallBooking__CoverageText'>
+                {label}
+              </Span>
+            </HStack>
+          ))}
+        </VStack>
+      </VStack>
+    </>
+  )
 }
 
 export const CallBooking = ({ callBooking, onBookCall }: CallBookingProps) => {
   const { t } = useTranslation()
   const { formatDate } = useIntlFormatter()
+  const countdownLabel = useCallBookingCountdownLabel(callBooking?.eventStartAt)
 
   if (callBooking == null) {
     return (
@@ -86,35 +104,7 @@ export const CallBooking = ({ callBooking, onBookCall }: CallBookingProps) => {
   const subtitle = t('callBookings:label.meet_bookkeeping_team', 'Meet with our bookkeeping team')
   const callPlatform = callBooking.callType === CallBookingType.ZOOM ? 'Zoom' : 'Google Meet'
   const callLink = callBooking.callLink.toString()
-  const countdownDescriptor = getCountdownDescriptor(Date.now(), callBooking.eventStartAt)
-  const countdownLabel = (() => {
-    switch (countdownDescriptor.kind) {
-      case 'days':
-        return tPlural(t, 'callBookings:state.call_in_count_days', {
-          count: countdownDescriptor.value,
-          displayCount: countdownDescriptor.value,
-          one: 'in {{displayCount}} day',
-          other: 'in {{displayCount}} days',
-        })
-      case 'hours':
-        return tPlural(t, 'callBookings:state.call_in_count_hours', {
-          count: countdownDescriptor.value,
-          displayCount: countdownDescriptor.value,
-          one: 'in {{displayCount}} hour',
-          other: 'in {{displayCount}} hours',
-        })
-      case 'startingSoon':
-        return t('callBookings:state.starting_soon', 'starting soon')
-      case 'none':
-        return ''
-    }
-  })()
   const timeLabel = formatDate(callBooking.eventStartAt, DateFormat.MonthDayWithTimeReadable)
-  const whatWeWillCoverItems: Array<[string, string]> = [
-    ['business_and_books', t('callBookings:label.cover_business_and_books', 'Walk through your business and books')],
-    ['accounts_and_documents', t('callBookings:label.cover_accounts_and_documents', 'Connect your accounts and documents')],
-    ['first_month_expectations', t('callBookings:label.cover_first_month_expectations', 'Set expectations for our first month')],
-  ]
 
   return (
     <Container name='CallBooking'>
@@ -157,41 +147,7 @@ export const CallBooking = ({ callBooking, onBookCall }: CallBookingProps) => {
           </Span>
         </HStack>
 
-        {isOnboardingCall && (
-          <>
-            <Span nonAria className='Layer__CallBooking__Divider' />
-
-            <VStack pbe='md'>
-              <Span
-                nonAria
-                size='2xs'
-                variant='subtle'
-                pbe='sm'
-                className='Layer__CallBooking__CoverageEyebrow'
-              >
-                {t('callBookings:label.what_well_cover', 'What we\'ll cover')}
-              </Span>
-              <VStack role='list' gap='xs'>
-                {whatWeWillCoverItems.map(([key, label]) => (
-                  <HStack
-                    key={key}
-                    className='Layer__CallBooking__CoverageItem'
-                    align='start'
-                    gap='sm'
-                    role='listitem'
-                  >
-                    <Span nonAria className='Layer__CallBooking__CoverageBadge'>
-                      <Check size={12} strokeWidth={2.5} />
-                    </Span>
-                    <Span size='sm' className='Layer__CallBooking__CoverageText'>
-                      {label}
-                    </Span>
-                  </HStack>
-                ))}
-              </VStack>
-            </VStack>
-          </>
-        )}
+        {isOnboardingCall && <OnboardingCallCoverage />}
 
         <HStack className='Layer__CallBooking__ActionRow' align='center' gap='xs'>
           <HStack className='Layer__CallBooking__JoinAction'>
